requirements.txt
numpy
scipy
matplotlib
torch
pyro-ppl
pycbc